Say what a craft knows, and where the sky's blind spots come from
22-provenance.md is the document the rest of these assume without stating: nobody has a view
of the world, only a pile of measurements with a witness, a time and a route on each. It
covers what a record is, why distance is the whole mechanic and what it costs to measure,
where a sky's blind spots come from — one scatter constant gives arcseconds around a
neighbouring star and tens of degrees around the sun the telescope is orbiting — and what is
deliberately not built yet, in the order it should be.

The telescope panel moves to its own module and the duty machinery out of `session` into
`watch`, both of which had crossed the thousand-line cap. `SelectNearest` picks from what has
been detected rather than from the catalogue, which is the same bug the map had.

+++ crates/lc-world/src/knowledge/survey.rs
@@ -118,7 +118,7 @@ pub fn hidden_by(sky: &[Source], index: usize, resolution_rad: f64) -> Option<St
         .enumerate()
         .filter(|(i, s)| *i != index && s.flux_w_m2 > target.flux_w_m2)
         .find(|(_, s)| {
-            let separation = s.toward.angle_between(target.toward) as f64;
+            let separation = s.toward.angle_between(target.toward);
             separation < glare_radius_rad(resolution_rad, s.flux_w_m2, target.flux_w_m2)
         })
         .map(|(_, s)| s.star)
@@ -511,7 +511,7 @@ mod tests {
         let sky = [source(1, truth, flux)];
         let seen = look(&optics, &sky, 0, 60.0, DVec3::ZERO, 1.0e6, Witness(1)).unwrap();
         assert_eq!(seen.band, Band::V);
-        let error = seen.bearing.toward.angle_between(truth) as f64;
+        let error = seen.bearing.toward.angle_between(truth);
         assert!(
             error < 5.0 * seen.bearing.sigma_rad,
             "{error} vs {}",
